BLE Mesh： Add the vendors morenode code. 1. Change the nvds driver and fix some Compatibility and stability issues; 2. Change the app code for fix the app key id issues; 3. Fix the warning issues; 4. Change the SVC heap from 4000 bytes to 5128 bytes; 5. Add the uart2 driver to the project.

+++ sdk/plactform/src/driver/uart/uart.h
@@ -71,7 +71,7 @@ void dbg_initial(void);
 void uart_clear_rxfifo(void);
 
 uint8_t Uart_Read_Byte(void);
-uint8_t read_uart_ringbuf_data(uint8_t *buf, uint16_t len);
+uint16_t read_uart_ringbuf_data(uint8_t *buf, uint16_t len);
 uint16_t Read_Uart_Buf(uint8_t *buf, uint16_t len);
